Add files via upload
needed some assistance from answer key, but basically the idea here was that we are tokenizing words from various article headlines, and creating a neural network to be able to categorize the headlines in a certain group. This involved training the network to assign vectors to certain meanings of words and be able to categorize these tokenized words based on the similarity of the meaning of the vector, thus making a network of similar words combined to categorize the headlines.