Add files via upload
This exercise involved creating a neural network that would provide meaning and sense to words in a sentence, rather than just jumbling words together without any sort of meaning. The idea behind this is that before, we had created a network that could take text and resemble then as tokens, but the tokens are just a label within a sentence; it could mean anything. We are trying to make sense of the sentences with the computer and use training data in order to understand that certain words have significant meaning.